requirements.txt
asgiref==3.7.2
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==23.2
sqlparse==0.4.4
whitenoise==6.6.0
psycopg2==2.9.9